put "expression-evaluator" folder in include folder and move contents
